% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{get_buildings}
\alias{get_buildings}
\title{Return tibble of data from file names}
\source{
\url{https://github.com/STAT210-S23/STAT210-Holyoke-History-Room/tree/main/Riverview_Renewal_Project}
}
\usage{
get_buildings()
}
\description{
Generate data on all buildings in the urban renewal project
}
\details{
Returns a tibble with all image information (and some additional attributes) for the Holyoke Urban Renewal project
}
\examples{
get_buildings()

}
