% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{get_filename}
\alias{get_filename}
\title{Get building filename from Holyoke Urban Renewal dataset}
\usage{
get_filename(formula)
}
\arguments{
\item{formula}{the formula describing the relationship}
}
\description{
XX
}
\details{
XX
}
\examples{
get_filename()

}
